issue 29, works fine

+++ src/main/java/org/pethotel/HeavenForPets/domein/Pet/Pet.java
@@ -21,7 +21,7 @@ public abstract class Pet implements Serializable {
     private long id;
     private String name;
     private String comment;
-    private long roomNumber;
+    private int roomNumber;
     private Date dateIn;
     private Date dateOut;
 
@@ -35,11 +35,11 @@ public abstract class Pet implements Serializable {
         this.id = id;
     }
 
-    public long getRoomNumber() {
+    public int getRoomNumber() {
         return roomNumber;
     }
 
-    public void setRoomNumber(long roomNumber) {
+    public void setRoomNumber(int roomNumber) {
         this.roomNumber = roomNumber;
     }
